Refactor: Move verify email logic to separate controller and update route to /api/verify-email
