+++ assets/js/pages/VerifyEmail.jsx
@@ -14,7 +14,7 @@ export default function VerifyEmail() {
             }
 
             try {
-                const res = await fetch(`/api/verify/email`, {
+                const res = await fetch(`/api/verify-email`, {
                     method: "POST",
                     headers: { "Content-Type": "application/json" },
                     body: JSON.stringify({ token }),
