+++ src/Controller/RegistrationController.php
@@ -93,50 +93,5 @@ class RegistrationController extends AbstractController
         ], Response::HTTP_CREATED);
     }
 
-    /**
-     * Handles email verification API endpoint
-     *
-     * @param Request $request
-     * @param UserEmailVerificationHandler $handler
-     * @return JsonResponse
-     */
-    #[Route('/api/verify/email', name: 'app_verify_email', methods: ['POST'])]
-    public function verifyEmail(
-        Request $request,
-        UserEmailVerificationHandler $handler,
-    ): JsonResponse
-    {
-        $data = json_decode($request->getContent());
-        $token = $data->token ?? null;
-
-        if (null === $token) {
-            return new JsonResponse([
-                'status' => 'error',
-                'code' => 'MISSING_TOKEN',
-                'message' => 'Missing verification token.'
-            ], Response::HTTP_BAD_REQUEST);
-        }
-
-        try {
-            $handler->handle($token);
-        } catch (\Exception $e) {
-            $this->logger->error('Email verification failed', [
-                'exception' => $e,
-                'message' => $e->getMessage(),
-            ]);
-
-            return new JsonResponse([
-                'status' => 'error',
-                'code' => 'VERIFICATION_FAILED',
-                'message' => 'Email verification failed.'
-            ], Response::HTTP_BAD_REQUEST);
-        }
-
-        return new JsonResponse([
-            'status' => 'success',
-            'message' => 'Email verified successfully.'
-        ]);
-    }
-
 }
